[config] move debug_interface_config to own file
per title

Closes: #1722
Approved by: mimoo
